PostJam version

+++ _Progect/Code/MapEntites/Mountain.cs
@@ -2,7 +2,7 @@ namespace Code.MapEntities
 {
     public class Mountain : ResourceEntity
     {
-        protected override void AddResource(int count)
+        protected override void AddResource(float count)
         {
             GameManager.Instance.AddRock(count);
         }
