+++ _Progect/Code/MapEntites/Wheat.cs
@@ -2,7 +2,7 @@ namespace Code.MapEntities
 {
     public class Wheat : ResourceEntity
     {
-        protected override void AddResource(int count)
+        protected override void AddResource(float count)
         {
             GameManager.Instance.AddWheat(count);
         }
